+++ room-states-repositories/src/test/kotlin/info/metadude/kotlin/library/roomstates/repositories/simple/SimpleRoomStatesRepositoryTest.kt
@@ -65,7 +65,7 @@ class SimpleRoomStatesRepositoryTest {
         return SimpleRoomStatesRepository(
             url = "https://example.com",
             path = "rooms",
-            httpClient = httpClient,
+            callFactory = httpClient,
             api = api,
         )
     }